^R Extract workflow validators into internal/metadatautil/workflows
First sub-PR of the metadatautil/validate.go decomposition (Section C
of the /sethify-remediation plan). Moves the GitHub Actions workflow
validation cluster from internal/metadatautil/validate.go into a new
internal/metadatautil/workflows/ subpackage.

Moved (with exported names so existing tests in package metadatautil
can keep using them via workflows.X):
- CollectWorkflowJobNames (was collectWorkflowJobNames)
- CheckWorkflows, EnsureWorkflowTools
- ValidateCodeQLWorkflow, ValidateReleaseWorkflowCodeQLFlow
- ValidateCIWorkflowPRShapeFlow
- ValidateReleaseWorkflowControlPlaneFlow
- ValidateReleaseWorkflowGitHubAttestationFlow
- ValidateMacOSInstallVerificationFlow
- ValidateUpstreamRefreshWorkflow
- workflowDocument, workflowJob types

The new subpackage avoids a circular dep with metadatautil by
inlining readText, requireStringSliceTable, and mustStringSlice
and calling tomlsubset.Parse directly instead of routing through
metadatautil.ParseTOMLSubset.

CheckPinnedInputs in metadatautil/validate.go now calls the new
workflows package for the CI / release / macOS install / upstream
refresh validators it invokes mid-pinning-validation.

isSafePullRequestTargetWorkflow + parseWorkflowRoot + yamlMappingValues
+ yamlScalarValue stay in validate.go for now; they move with
CheckPinnedInputs when that cluster itself is extracted in a later
sub-PR. Likewise, the hosted-controls validators and tests stay in
metadatautil for the hostedcontrols sub-PR.

Drops 341 LOC from validate.go. go build/test/gofmt clean; no
behavior change.
